refactor: 사용자 정의 예외처리 작업
