+++ src/main/java/com/recipesns/post/converter/JsonToPostImagesConverter.java
@@ -1,6 +1,7 @@
 package com.recipesns.post.converter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.recipesns.post.converter.exception.PostImagesConversionException;
 import com.recipesns.post.domain.PostImages;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +22,7 @@ public class JsonToPostImagesConverter implements Converter<byte[], PostImages> 
             String jsonSource = objectMapper.readValue(source, String.class);
             return objectMapper.readValue(jsonSource, PostImages.class);
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            throw new PostImagesConversionException("JSON 문자열을 이미지 리스트로 변환할 수 없습니다", e);
         }
     }
 }
