+++ src/main/java/com/recipesns/post/converter/PostImagesToJsonConverter.java
@@ -2,6 +2,7 @@ package com.recipesns.post.converter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.recipesns.post.converter.exception.JsonConversionException;
 import com.recipesns.post.domain.PostImages;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,7 +20,7 @@ public class PostImagesToJsonConverter implements Converter<PostImages, String> 
         try {
             return objectMapper.writeValueAsString(images);
         } catch (JsonProcessingException e) {
-            throw new RuntimeException("Could not convert list of PostImage to JSON string", e);
+            throw new JsonConversionException("이미지 리스트를 JSON 문자열로 변환할 수 없습니다", e);
         }
     }
 }
