+++ src/main/java/com/recipesns/post/converter/exception/JsonConversionException.java
@@ -1,0 +1,9 @@
+package com.recipesns.post.converter.exception;
+
+public class JsonConversionException extends RuntimeException {
+
+    public JsonConversionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}
