+++ src/main/java/com/recipesns/post/converter/exception/PostImagesConversionException.java
@@ -1,0 +1,9 @@
+package com.recipesns.post.converter.exception;
+
+public class PostImagesConversionException extends RuntimeException {
+
+    public PostImagesConversionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}
